# docs/swagger.yaml
definitions:
...
  github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.ListPostsPrivateRow:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      subtitle:
        type: string
      tags: {}
      title:
        type: string
    type: object
...
  github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post:
    properties:
      category_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      publicated:
        type: boolean
      subtitle:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.PostsTag:
    properties:
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      tag_id:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_api.createPostRequest:
    properties:
      category_id:
        type: string
      content:
        type: string
      publicated:
        type: boolean
      subtitle:
        type: string
      title:
        type: string
    required:
    - category_id
    - content
    - subtitle
    - title
    type: object
  internal_api.createPostTag:
    properties:
      post_id:
        type: string
      tag_id:
        type: string
    required:
    - post_id
    - tag_id
    type: object
...
  internal_api.updatePostRequest:
    properties:
      category_id:
        type: string
      content:
        $ref: '#/definitions/pgtype.Text'
      publicated:
        $ref: '#/definitions/pgtype.Bool'
      subtitle:
        $ref: '#/definitions/pgtype.Text'
      title:
        $ref: '#/definitions/pgtype.Text'
    type: object
...
  pgtype.Bool:
    properties:
      bool:
        type: boolean
      valid:
        type: boolean
    type: object
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /admin/category-post/{id}:
    get:
      consumes:
      - application/json
      description: Recive the one post on the admin panel by Category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      security:
      - JWT: []
      summary: Get a Post by Category Private
      tags:
      - post
      - list
  /admin/post:
    post:
      consumes:
      - application/json
      description: Create a new Post
      parameters:
      - description: post Data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/internal_api.createPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      security:
      - JWT: []
      summary: Create a new Post
      tags:
      - post
      - create
  /admin/post-tag:
    post:
      consumes:
      - application/json
      description: Create a new PostTag
      parameters:
      - description: post_tag Data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/internal_api.createPostTag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.PostsTag'
      security:
      - JWT: []
      summary: Create a new PostTag
      tags:
      - post_tag
      - create
  /admin/post-tag/{id}:
    delete:
      consumes:
      - application/json
      description: Delete one post_tag on the admin panel
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.PostsTag'
      security:
      - JWT: []
      summary: Delete a PostTag by Id
      tags:
      - post_tag
      - delete
  /admin/post/{id}:
    delete:
      consumes:
      - application/json
      description: Delete one post on the admin panel
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      security:
      - JWT: []
      summary: Delete a Post by Id
      tags:
      - post
      - delete
    get:
      consumes:
      - application/json
      description: Recive the one post on the admin panel
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      security:
      - JWT: []
      summary: Get a Post by Id Private
      tags:
      - post
      - get
    put:
      consumes:
      - application/json
      description: Update a new Post
      parameters:
      - description: post Data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/internal_api.updatePostRequest'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      security:
      - JWT: []
      summary: Update a Post
      tags:
      - post
      - update
  /admin/posts:
    get:
      description: Recive all posts on the admin panel
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.ListPostsPrivateRow'
      security:
      - JWT: []
      summary: List all Posts Private
      tags:
      - post
      - list
  /admin/tag-post/{id}:
    get:
      consumes:
      - application/json
      description: Recive the one post on the admin panel by Tag
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      security:
      - JWT: []
      summary: Get a Post by tag Private
      tags:
      - post
      - list
...
  /category-post/{id}:
    get:
      consumes:
      - application/json
      description: Recive the one post by Category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      summary: Get a Post by Category Public
      tags:
      - post
      - list
...
  /post/{id}:
    get:
      consumes:
      - application/json
      description: Recive the one post public
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      summary: Get a Post by Id Public
      tags:
      - post
      - get
...
  /tag-post/{id}:
    get:
      consumes:
      - application/json
      description: Recive the one post by Tag
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JairoRiver_personal_blog_backend_internal_db_sqlc.Post'
      summary: Get a Post by tag Public
      tags:
      - post
      - list
...
securityDefinitions:
  JWT:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"